Move shard specific logic into shard.py
A Shard owns ShardState, Miner, TransactionGenerator, Synchronizer,
and PeerShardConnection for each cluster.
Functions manipulating a single shard should stay in shard.py.

Slave owns a list of Shards, SlaveConnectionManager,
and MasterConnection.
Functions for communicating with slaves, master should stay
in slave.py including outgoing RPC requests and incoming RPC handlers.
