+++ quarkchain/cluster/tests/test_utils.py
@@ -1,22 +1,22 @@
 import asyncio
 from contextlib import ContextDecorator
 
-from quarkchain.cluster.master import MasterServer
-from quarkchain.cluster.root_state import RootState
-from quarkchain.cluster.simple_network import SimpleNetwork
-from quarkchain.cluster.slave import SlaveServer
 from quarkchain.cluster.cluster_config import (
     ClusterConfig,
     SlaveConfig,
     SimpleNetworkConfig,
 )
-from quarkchain.env import DEFAULT_ENV
-from quarkchain.core import Address, Transaction, Code, ShardMask
+from quarkchain.cluster.master import MasterServer
+from quarkchain.cluster.root_state import RootState
+from quarkchain.cluster.simple_network import SimpleNetwork
+from quarkchain.cluster.slave import SlaveServer
+from quarkchain.core import Address, Branch, Transaction, Code, ShardMask
 from quarkchain.db import InMemoryDb
+from quarkchain.env import DEFAULT_ENV
 from quarkchain.evm.transactions import Transaction as EvmTransaction
+from quarkchain.genesis import GenesisManager
 from quarkchain.protocol import AbstractConnection
 from quarkchain.utils import call_async, check
-from quarkchain.genesis import GenesisManager
 
 
 def get_test_env(
@@ -36,7 +36,9 @@ def get_test_env(
 
     # fund genesis account in all shards
     for i, shard in enumerate(env.quark_chain_config.SHARD_LIST):
-        shard.GENESIS.ALLOC[genesis_account.address_in_shard(i).serialize().hex()] = genesis_minor_quarkash
+        shard.GENESIS.ALLOC[
+            genesis_account.address_in_shard(i).serialize().hex()
+        ] = genesis_minor_quarkash
 
     GenesisManager.finalize_config(env.quark_chain_config)
 
@@ -156,12 +158,18 @@ class Cluster:
         self.network = network
         self.peer = peer
 
-    def get_shard_state(self, shard_id):
-        branch_value = self.master.env.quark_chain_config.SHARD_SIZE | shard_id
+    def get_shard(self, shard_id):
+        branch = Branch.create(self.master.env.quark_chain_config.SHARD_SIZE, shard_id)
         for slave in self.slave_list:
-            if branch_value in slave.shard_state_map:
-                return slave.shard_state_map[branch_value]
+            if branch in slave.shards:
+                return slave.shards[branch]
         return None
+
+    def get_shard_state(self, shard_id):
+        shard = self.get_shard(shard_id)
+        if not shard:
+            return None
+        return shard.state
 
 
 # server.close() does not release the port sometimes even after server.wait_closed() is awaited.
